Fix a number of issues:
1. Not all Amber topologies must have bonds/angles/dihedrals with and without
hydrogen (take pure water, for instance, which has only bonds including
hydrogen). This used to cause a crash because those sections couldn't be found
in the parsed prmtop.  This was fixed and a test was added (with pure water).

2. Add a UnitCell attribute (along with associated getter and setter) for the
AmberParm object, that is set upon parsing from the BOX_DIMENSIONS section. Add
a test for this as well.

3. Add the missing RADII and SCREEN sections to the 4096wat.parm7 prmtop file,
since the parser here is not flexible enough to deal without it (this only
affects converted old-style prmtops, since RADII and SCREEN are always printed
by default).

4. Add a missing define in exceptions.h which was always masked because <string>
was included everywhere "exceptions.h" was *before* exceptions.h was included.
As a result, <string> was always available accidentally. This wasn't true in
unitcell.cpp, so we finally got an error exposing this bug.